Modification pour afficher correctement les sauts de lignes pour la notice, infos et biographie auteur dans modifierannonce et visualiserAnnoce

+++ src/components/api/effaceAnnonce.php
@@ -10,33 +10,40 @@
     else {$id = htmlspecialchars(trim($_POST['id']));}
 
     $id = intval ($id);
+    $return = $id;
     
     include ("connexion.php");
-        
-    $req = "DELETE FROM images WHERE img_id=" . $id;
+
+    // On récupère la ref de l'annonce pour la suprimer sur le serveur
+    $req = "SELECT ref FROM images WHERE img_id = '".$id."'";
     $sth = $cnx->prepare($req);
-    $return = $sth->execute();	
-
-    echo"";
-
-    if ($return)
-	{
-        echo "L'annonce a été effacée en base\n";
-
-        // On récupère la ref de l'annonce pour la suprimer sur le serveur
-        $req = "SELECT ref FROM images WHERE img_id = '".$id."'";
+    $sth->execute();
+    $col = $sth->fetch(PDO::FETCH_BOTH);
+    $ref = $col[0];
+    
+    if ($ref!='')
+    {
+        $req = "DELETE FROM images WHERE img_id=" . $id;
         $sth = $cnx->prepare($req);
-        $sth->execute();
-        $col = $sth->fetch(PDO::FETCH_BOTH);
+        $return = $sth->execute();	
 
-        $ref = $col[0];
+        if ($return)
+        {
+            echo "L'annonce a été effacée en base\n";
 
-        // Supprime les images et le répertoire parent de Hostinger via FTP
-        include ("deletePicToFtp.php");	
-	}
-	else
-	{
-		echo "Un prôblème est survenue lors de la suppression de l'annonce en base\n";
+            
+
+            // Supprime les images et le répertoire parent de Hostinger via FTP
+            include ("deletePicToFtp.php");	
+        }
+        else
+        {
+            echo "Un prôblème est survenue lors de la suppression de l'annonce en base\n";
+        }
+    }
+    else
+    {
+        echo "Il n'y a pas de référence pour l'annonce avec l'id : $id\n";
     }
     
     echo json_encode($return);
